mv cuda to common backends
